Refactor Cython integration and improve class search functionality
