+++ setup.cfg
@@ -23,9 +23,9 @@ classifiers =
 	Programming Language :: Python :: 3.8
 	Topic :: Scientific/Engineering
 tests_require = 
-	pytest~=8.3.3
-	fmpy~=0.3.21
-    cython~=3.0.11
+	pytest
+	fmpy
+    cython
 
 [options]
 include_package_data = True
